backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0
sse-starlette>=1.6.0